fix(workflow-apps-defaults): make apps default read from appropriate home
Makes workflow-apps-defaults depend on the shared/env approach to finding the
workflow-home directory. This ensures that the apps are found when the
WORKFLOW_HOME enviroment variable is set, but also that refactorings of
shared/env will be reflected in the workflow-apps-defaults package.
