+++ rollup.config.js
@@ -9,13 +9,7 @@ import { resolve, join } from 'path';
 
 import flatMap from 'lodash.flatmap';
 
-const blackList = [
-  'workflow-apps-defaults',
-  'workflow-layout',
-  'workflow-layouts',
-  'workflow-template',
-  'workflow-web',
-];
+const blackList = ['workflow-layout', 'workflow-layouts', 'workflow-template', 'workflow-web'];
 
 const nodeInternalDependencies = {
   'create-workflow-app': ['path', 'util', 'os', 'child_process'],
